Update Color and Ray utilities

+++ src/Core/Ray.cpp
@@ -2,8 +2,8 @@
 
 #include <glm/vec3.hpp>
 
-
 namespace raytracer {
+
     Ray::Ray(const glm::vec3& origin, const glm::vec3& direction)
         : m_Origin{origin}, m_Direction{direction}
     {
@@ -29,9 +29,9 @@ namespace raytracer {
         return m_Direction;
     }
 
-    glm::vec3 Ray::At(const double t) const
+    glm::vec3 Ray::At(const float t) const
     {
         return m_Origin + t * m_Direction;
     }
-
+    
 }
